Refactoring ply loading based on format
