+++ src/index.js
@@ -1,5 +1,5 @@
 import { PlyParser } from './loaders/ply/PlyParser.js';
-import { CompressedPlyParser } from './loaders/ply/CompressedPlyParser.js';
+import { PlayCanvasCompressedPlyParser } from './loaders/ply/PlayCanvasCompressedPlyParser.js';
 import { PlyLoader } from './loaders/ply/PlyLoader.js';
 import { SplatLoader } from './loaders/splat/SplatLoader.js';
 import { KSplatLoader } from './loaders/ksplat/KSplatLoader.js';
@@ -20,7 +20,7 @@ import { SceneRevealMode } from './SceneRevealMode.js';
 
 export {
     PlyParser,
-    CompressedPlyParser,
+    PlayCanvasCompressedPlyParser,
     PlyLoader,
     SplatLoader,
     KSplatLoader,
