+++ src/loaders/ply/PlayCanvasCompressedPlyParser.js
@@ -69,7 +69,7 @@ const getElementPropStorage = (element, name) => {
     ?.storage;
 };
 
-export class CompressedPlyParser {
+export class PlayCanvasCompressedPlyParser {
 
   static decodeHeaderText(headerText) {
 
@@ -203,7 +203,7 @@ export class CompressedPlyParser {
       buf.slice(0, endHeaderTokenOffset)
     );
 
-    const {chunkElement, vertexElement, bytesPerSplat} = CompressedPlyParser.decodeHeaderText(headerText);
+    const {chunkElement, vertexElement, bytesPerSplat} = PlayCanvasCompressedPlyParser.decodeHeaderText(headerText);
 
     return {
       'headerSizeBytes': endHeaderTokenOffset + HeaderEndTokenBytes.length,
@@ -268,10 +268,11 @@ export class CompressedPlyParser {
 
   static readPly(plyBuffer, propertyFilter = null) {
 
-    const header = CompressedPlyParser.decodeHeader(plyBuffer);
+    const header = PlayCanvasCompressedPlyParser.decodeHeader(plyBuffer);
 
-    let readIndex = CompressedPlyParser.readElementData(header.chunkElement, plyBuffer, header.headerSizeBytes, null, null, propertyFilter);
-    CompressedPlyParser.readElementData(header.vertexElement, plyBuffer, readIndex, null, null, propertyFilter);
+    let readIndex = PlayCanvasCompressedPlyParser.readElementData(header.chunkElement, plyBuffer,
+                                                                  header.headerSizeBytes, null, null, propertyFilter);
+    PlayCanvasCompressedPlyParser.readElementData(header.vertexElement, plyBuffer, readIndex, null, null, propertyFilter);
 
     return {
       'chunkElement': header.chunkElement,
@@ -359,7 +360,7 @@ export class CompressedPlyParser {
   static parseToUncompressedSplatBufferSection(chunkElement, vertexElement, fromIndex, toIndex, chunkSplatIndexOffset,
                                                vertexDataBuffer, veretxReadOffset, outBuffer, outOffset, propertyFilter = null) {
 
-    CompressedPlyParser.readElementData(vertexElement, vertexDataBuffer, veretxReadOffset, fromIndex, toIndex, propertyFilter);
+    PlayCanvasCompressedPlyParser.readElementData(vertexElement, vertexDataBuffer, veretxReadOffset, fromIndex, toIndex, propertyFilter);
 
     const outBytesPerCenter = SplatBuffer.CompressionLevels[0].BytesPerCenter;
     const outBytesPerScale = SplatBuffer.CompressionLevels[0].BytesPerScale;
@@ -367,14 +368,14 @@ export class CompressedPlyParser {
     const outBytesPerSplat = SplatBuffer.CompressionLevels[0].SphericalHarmonicsDegrees[0].BytesPerSplat;
 
     const { positionExtremes, scaleExtremes, position, rotation, scale, color } =
-      CompressedPlyParser.getElementStorageArrays(chunkElement, vertexElement);
+      PlayCanvasCompressedPlyParser.getElementStorageArrays(chunkElement, vertexElement);
 
     const OFFSET = UncompressedSplatArray.OFFSET;
     const tempSplat = UncompressedSplatArray.createSplat();
 
     for (let i = fromIndex; i <= toIndex; ++i) {
 
-      CompressedPlyParser.decompressSplat(i, chunkSplatIndexOffset, position, positionExtremes,
+      PlayCanvasCompressedPlyParser.decompressSplat(i, chunkSplatIndexOffset, position, positionExtremes,
                                           scale, scaleExtremes, rotation, color, tempSplat);
 
       const outBase = i * outBytesPerSplat + outOffset;
@@ -404,19 +405,19 @@ export class CompressedPlyParser {
   }
 
   static parseToUncompressedSplatArray(plyBuffer) {
-    const { chunkElement, vertexElement } = CompressedPlyParser.readPly(plyBuffer);
+    const { chunkElement, vertexElement } = PlayCanvasCompressedPlyParser.readPly(plyBuffer);
 
     const splatArray = new UncompressedSplatArray();
 
     const { positionExtremes, scaleExtremes, position, rotation, scale, color } =
-      CompressedPlyParser.getElementStorageArrays(chunkElement, vertexElement);
+      PlayCanvasCompressedPlyParser.getElementStorageArrays(chunkElement, vertexElement);
 
     for (let i = 0; i < vertexElement.count; ++i) {
 
       splatArray.addDefaultSplat();
       const newSplat = splatArray.getSplat(splatArray.splatCount - 1);
 
-      CompressedPlyParser.decompressSplat(i, 0, position, positionExtremes, scale, scaleExtremes, rotation, color, newSplat);
+      PlayCanvasCompressedPlyParser.decompressSplat(i, 0, position, positionExtremes, scale, scaleExtremes, rotation, color, newSplat);
     }
 
     const mat = new THREE.Matrix4();
