+++ src/loaders/ply/PlyFormat.js
@@ -1,0 +1,5 @@
+export const PlyFormat = {
+    'INRIAV1': 0,
+    'INRIAV2': 1,
+    'PlayCanvasCompressed': 2
+};
